+++ tests/calculators/qchem/test_qchem.py
@@ -213,6 +213,4 @@ def test_qchem_read_results_freq(tmpdir):
     assert len(calc.results["qc_output"]["frequency_mode_vectors"]) == 36
     assert len(calc.results["qc_output"]["frequency_mode_vectors"][0]) == 14
     assert len(calc.results["qc_output"]["frequency_mode_vectors"][0][0]) == 3
-    assert calc.results["qc_output"]["enthalpy"] == 2.647248450819514
-    assert calc.results["qc_output"]["entropy"] == 0.003996739364205975
     tmpdir.chdir()